# shared/contracts/api/aevora-v1.openapi.yaml
openapi: 3.1.0
info:
  title: Aevora API
  version: v1
servers:
  - url: https://api.aevora.example.com
paths:
...
  /v1/admin/runtime-config/releases/promote:
    post:
      summary: Promote a validated remote-config candidate
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminPromoteRuntimeConfigRequest"
      responses:
        "201":
          description: Promoted remote-config release
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RuntimeConfigPromotionResponse"
...
  /v1/admin/content/releases/promote:
    post:
      summary: Promote a validated launch-content candidate
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminPromoteContentRequest"
      responses:
        "201":
          description: Promoted content release
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContentPromotionResponse"
...
  /v1/admin/assets/releases/promote:
    post:
      summary: Promote a validated asset manifest
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminPromoteAssetManifestRequest"
      responses:
        "201":
          description: Promoted asset release
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AssetPromotionResponse"
  /v1/admin/accounts/{userId}:
    get:
      summary: Fetch support-safe account state for an operator
      parameters:
        - in: path
          name: userId
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Support-safe account summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminAccountSummaryResponse"
  /v1/admin/accounts/{userId}/export:
    post:
      summary: Prepare a support-safe export for an operator
      parameters:
        - in: path
          name: userId
          required: true
          schema:
            type: string
      responses:
        "201":
          description: Export request summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AccountExportResponse"
...
components:
  schemas:
...
    AccountExportResponse:
      type: object
      required: [exportRequest, summary]
      properties:
        exportRequest:
          type: object
          required: [id, status, generatedAt, availableUntil, format, redactionProfile, includes]
          properties:
            id:
              type: string
            status:
              type: string
              enum: [prepared]
            generatedAt:
              type: string
              format: date-time
            availableUntil:
              type: string
              format: date-time
            format:
              type: string
              enum: [json]
            redactionProfile:
              type: string
            includes:
              type: array
              items:
                type: string
        summary:
          type: object
          required: [userId, authMode, activeVows, archivedVows, completionDayCount, rewardGrantCount, inventoryItemCount, subscriptionTier, billingState, lastCompletionLocalDate]
          properties:
            userId:
              type: string
            authMode:
              type: string
              enum: [guest, registered]
            activeVows:
              type: integer
            archivedVows:
              type: integer
            completionDayCount:
              type: integer
            rewardGrantCount:
              type: integer
            inventoryItemCount:
              type: integer
            subscriptionTier:
              type: string
              enum: [free, trial, premium_monthly, premium_annual]
            billingState:
              type: string
              enum: [active, grace_period, billing_retry, expired]
            lastCompletionLocalDate:
              type: [string, "null"]
...
    RuntimeConfigReleaseRecord:
      type: object
      required: [id, label, sourcePath, schemaVersion, etag, promotedAt, promotedBy]
      properties:
        id:
          type: string
        label:
          type: string
        sourcePath:
          type: string
        schemaVersion:
          type: string
        etag:
          type: string
        promotedAt:
          type: string
          format: date-time
        promotedBy:
          type: string
        notes:
          type: [string, "null"]
...
    AdminPromoteRuntimeConfigRequest:
      type: object
      required: [candidatePath, label]
      properties:
        candidatePath:
          type: string
        label:
          type: string
        notes:
          type: [string, "null"]
    RuntimeConfigPromotionResponse:
      type: object
      required: [release, activeRelease]
      properties:
        release:
          $ref: "#/components/schemas/RuntimeConfigReleaseRecord"
        activeRelease:
          $ref: "#/components/schemas/RuntimeConfigReleaseRecord"
    ContentReleaseRecord:
      type: object
      required: [id, label, contentPath, copyPath, contentVersion, copyVersion, etag, promotedAt, promotedBy]
      properties:
        id:
          type: string
        label:
          type: string
        contentPath:
          type: string
        copyPath:
          type: string
        contentVersion:
          type: string
        copyVersion:
          type: string
        etag:
          type: string
        promotedAt:
          type: string
          format: date-time
        promotedBy:
          type: string
        notes:
          type: [string, "null"]
...
    AdminPromoteContentRequest:
      type: object
      required: [contentPath, label]
      properties:
        contentPath:
          type: string
        copyPath:
          type: [string, "null"]
        label:
          type: string
        notes:
          type: [string, "null"]
    ContentPromotionResponse:
      type: object
      required: [release, activeRelease]
      properties:
        release:
          $ref: "#/components/schemas/ContentReleaseRecord"
        activeRelease:
          $ref: "#/components/schemas/ContentReleaseRecord"
    AssetReleaseRecord:
      type: object
      required: [id, label, manifestPath, schemaVersion, releaseId, assetCount, etag, promotedAt, promotedBy]
      properties:
        id:
          type: string
        label:
          type: string
        manifestPath:
          type: string
        schemaVersion:
          type: string
        releaseId:
          type: string
        assetCount:
          type: integer
        etag:
          type: string
        promotedAt:
          type: string
          format: date-time
        promotedBy:
          type: string
        notes:
          type: [string, "null"]
...
    AdminPromoteAssetManifestRequest:
      type: object
      required: [manifestPath, label]
      properties:
        manifestPath:
          type: string
        label:
          type: string
        notes:
          type: [string, "null"]
    AssetPromotionResponse:
      type: object
      required: [release, activeRelease]
      properties:
        release:
          $ref: "#/components/schemas/AssetReleaseRecord"
        activeRelease:
          $ref: "#/components/schemas/AssetReleaseRecord"
    AdminAccountSummaryResponse:
      type: object
      required: [account]
      properties:
        account:
          type: object
          required: [userId, authMode, selectedIdentityShellId, originFamilyId, activeVowCount, archivedVowCount, completionDayCount, lastCompletionLocalDate, inventoryItemCount, rewardGrantCount, chapterId, chapterDay, districtStage, subscriptionTier, billingState, sourceConnections]
          properties:
            userId:
              type: string
            authMode:
              type: string
              enum: [guest, registered]
            selectedIdentityShellId:
              type: string
            originFamilyId:
              type: string
            activeVowCount:
              type: integer
            archivedVowCount:
              type: integer
            completionDayCount:
              type: integer
            lastCompletionLocalDate:
              type: [string, "null"]
            inventoryItemCount:
              type: integer
            rewardGrantCount:
              type: integer
            chapterId:
              type: string
            chapterDay:
              type: integer
            districtStage:
              type: string
            subscriptionTier:
              type: string
              enum: [free, trial, premium_monthly, premium_annual]
            billingState:
              type: string
              enum: [active, grace_period, billing_retry, expired]
            sourceConnections:
              type: array
              items:
                type: object
                required: [sourceType, authorizationState, supportedDomains]
                properties:
                  sourceType:
                    type: string
                    enum: [healthkit]
                  authorizationState:
                    type: string
                    enum: [not_requested, requested, granted, denied]
                  supportedDomains:
                    type: array
                    items:
                      type: string
                      enum: [workout, steps, sleep]
...